DDF-4185 Moved ServiceProperties into catalog-ui-search-api bundle
