+++ catalog/ui/catalog-ui-search-api/src/main/java/org/codice/ddf/catalog/ui/splitter/ServiceProperties.java
@@ -11,7 +11,7 @@
  * License is distributed along with this program and can be found at
  * <http://www.gnu.org/licenses/lgpl.html>.
  */
-package org.codice.ddf.catalog.ui.metacard.internal;
+package org.codice.ddf.catalog.ui.splitter;
 
 import java.util.Map;
 import java.util.Set;
